Add files via upload
Since the original text file, is too large to upload one can find it on: https://ride.citibikenyc.com/system-data or you can refer to my CSV file which contains the same content. The main.py file is the usage of an API (PTV), as one can see, the API key is not included for obvious reasons. Lastly, the SQL file includes all of my code, the table commands are not included because I used MYSQL workbench, which gives me the option to make tables by right-clicking on a specific schema, which will then allow us to create a table that way. Other than that, to have a better idea of what is going on within this code, one can review the pdf file which is the assignment itself.